requirements.txt
Flask>=2.3.3,<3.0
requests==2.31.0
networkx==3.1
python-dotenv==1.0.0
